pyproject.toml
[project]
name = "liked-videos"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Michael Uloth", email = "hello@michaeluloth.com" }
]
dependencies = [
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.137.0",
    "mypy>=1.10.1",
    "pydantic>=2.8.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"liked-videos" = "liked_videos:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/liked_videos"]
